Move voice to separate file
